streamlit_app_v.0.1/requirements.txt
streamlit
plotly
Pillow
pandas
os